+++ src/main/java/com/ticketresolve/ticketresolve/configuration/security/jwt/JwtUtils.java
@@ -1,4 +1,4 @@
-package com.ticketresolve.ticketresolve.security.jwt;
+package com.ticketresolve.ticketresolve.configuration.security.jwt;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
