--- src/main/java/com/ticketresolve/ticketresolve/controller/LoginRequest.java
@@ -1,9 +1,0 @@
-package com.ticketresolve.ticketresolve.controller;
-
-import lombok.Data;
-
-@Data
-public class LoginRequest {
-     private String username;
-    private String password;
-}
