--- src/main/java/com/ticketresolve/ticketresolve/model/ERole.java
@@ -1,7 +1,0 @@
-package com.ticketresolve.ticketresolve.model;
-
-public enum ERole {
-    ADMINISTRADOR,
-    CLIENTE,
-    EMPLEADO,
-}
