+++ src/main/java/com/ticketresolve/ticketresolve/persistence/entity/ERole.java
@@ -1,0 +1,6 @@
+package com.ticketresolve.ticketresolve.persistence.entity;
+
+public enum ERole {
+    ADMINISTRADOR,
+    EMPLEADO,
+}
