+++ src/main/java/com/ticketresolve/ticketresolve/persistence/repository/IUsuarioRepository.java
@@ -1,15 +1,15 @@
-package com.ticketresolve.ticketresolve.repository;
+package com.ticketresolve.ticketresolve.persistence.repository;
 
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
-import com.ticketresolve.ticketresolve.model.Usuario;
+import com.ticketresolve.ticketresolve.persistence.entity.Usuario;
 
 import java.util.Optional;
 
 @Repository
-public interface UsuarioRepository extends CrudRepository<Usuario, Long> {
+public interface IUsuarioRepository extends CrudRepository<Usuario, Long> {
 
     Optional<Usuario> findByUsername(String username);
