+++ src/main/java/com/ticketresolve/ticketresolve/presentation/dto/LoginRequest.java
@@ -1,0 +1,11 @@
+package com.ticketresolve.ticketresolve.presentation.dto;
+
+import lombok.Data;
+
+@Data
+public class LoginRequest {
+
+    private String username;
+    private String password;
+
+}
