--- src/main/java/com/ticketresolve/ticketresolve/repository/RoleRepository.java
@@ -1,11 +1,0 @@
-package com.ticketresolve.ticketresolve.repository;
-
-import org.springframework.data.repository.CrudRepository;
-import org.springframework.stereotype.Repository;
-
-import com.ticketresolve.ticketresolve.model.Role;
-
-@Repository
-public interface RoleRepository extends CrudRepository<Role, Long>{
-
-}
